added nstep estimation of target
